requirements.txt
supermemory
python-dotenv>=1.0.0
